Allow reusing image requested with lower min detail
When an image is requested to be loaded with different params
(imageParams_t), but the only difference is that the first one had a
larger minDimension, reuse the image. This happens with the
models/buildables/arm/screen_s specular image, which is used with two
diffuse maps that want different imageMinDimension.