Make R_RemapShader multiple registration-aware
Make R_RemapShader handle shader registration flags correctly. For each
version of the shader, register the replacement shader with a matching
set of flags, instead of always using RSF_DEFAULT.

Also this fixes a bug that when the replacement shader was not found,
the 'defaulted' shader (one that was not found or failed parsing) was
used for rendering. New behavior is to not change the replacement state.
Note that Thunder has recently been relying on this bug - see
https://github.com/UnvanquishedAssets/map-thunder_src.dpkdir/pull/7