Correctly set glConfig.relectionMappingAvailable
glConfig.reflectionMappingAvailable was set after shaders are loaded, so
gl_reflectionShader was not loaded when it should be. Set it in
EnableAvailableFeatures().

Also it could only be set to true; if you turned off the cvar the next
map change would not set it to false.

Fixes the bug that r_showCubeProbes causes a crash (due to null
gl_reflectionShader) if reflection mapping was not enabled in the
previous renderer load.