external_deps: make pkg-config working when cross-compiling
Without setting PKG_CONFIG_PATH to the cross database, pkg-config will only
report libraries installed for the native target, not for the cross target.

SDL3 relies on pkg-config to enable Wayland support, so we need that to
enable Wayland support when cross-compiling SDL3 when the related libraries
are only installed for the cross target.