Make global fog independent of view direction
For calculating the distance in fog in the global fog shader, only the
distance along the viewing axis was counted. This made it so there is
less fog when look at something from the corner of your eye than when
you look at it directly. This behavior matches fogQuake3 but there is
really no reason to do it. Count the whole distance instead of only
distance along view axis.