Fog: count full distance, not only along view axis
When calculating the distance from the viewer to a vertex in the
fogQuake3 shader, use the full distance, instead of dotting the
viewer-to-vertex vector with the view axis. So now all points on the
screen will be rendered consistently, instead of for example a point 30
degrees away from the center having its fog distance multiplied by
sqrt(2)/2 ~= 0.7.