Refactor: treat shader deform index like a GLSL macro
Instead of passing the deform index as a BindProgram argument, use a
SetDeform function that stores the deform index in the shader object.
This makes it work the same way as setting GLSL macro options (e.g.
SetVertexSkining), which makes sense because the deform index is also an
option that determines which variant of the GLSL shader gets compiled.
SetDeform is a member of GLDeformStage, so you are no longer allowed
and no longer required to pass a deform index to GLSL shaders that don't
use deforms.

This eliminates a fair amount of plumbing of the deform index through
functions that shouldn't have to care about it.