NUKE fog image but try to emulate it exactly in GLSL
I want to NUKE the fog image and replace it with a simple calculation in
GLSL. But first, to make sure I'm not making a mistake somewhere, I
wanted to try to replicate the old behavior as exactly as possible.
Turns out I had to emulate texture filtering to get it reasonably close;
just using FogFactor with the actual s and t values resulted in
substantial differences in brightness.

So this commit NUKES the fog image but replicates it with overly
complicated GLSL.

Fixes the artifacts where the fog plane intersects a model when viewed
from inside the fog.