Use a piecewise function for distance -> fog alpha
The original Quake 3 fog code in principle used the equation alpha =
sqrt(density * distance in fog). But the values were looked up in an
image that had a resolution that for small distances was way too low to
approximate the function well. Actually using sqrt at the low range is
bad because sqrt has an infinite at zero. Use a linear ramp for the
first bit which prevents hard edges caused by using pure sqrt and gives
a similar result to the old code.

This fixes color banding caused by the discrete fog texture. Also, it
makes the thickness of the fog appear the same regardless of how far you
are away from it.