Refactor glTextureBarrier code into function
Since it's about to be used more times. I named it like
PrepareForSamplingDepthMap instead of something about texture barrier,
because I may want to add more stuff in for #1814 where there is the
possibility to use a copy of the depth texture instead.

Also, always mark depth buffer dirtiness with material system.