external_deps: add the linux-riscv64-default profile
Makes it possible to build the deps for a linux-riscv64 build,
it's only useful for a server as NaCl isn't working.

It's used as an example of how we can implement and
unlock systems we don't support.

Such systems are meant to be listed as “extra”.

In that context, “supported” means we provide a release binary
in the unizip, while “extra” means we don't prevent
someone to build it and run it if that works for him.