Don't set CMAKE_PREFIX_PATH to external deps dir
We were adding the external deps location to both CMAKE_FIND_ROOT_PATH
and CMAKE_PREFIX_PATH. Only one is needed. CMAKE_PREFIX_PATH has the
weird behavior that if CMAKE_FIND_ROOT_PATH and CMAKE_PREFIX_PATH are
both set, each (root path) X (prefix path) pair is considered, and
when one isn't a subpath of the other, the search directory that
results is formed by concatenating the root path and the prefix path.
The deps dir is always an absolute path and not something that we want
to be concatenated in the middle, so avoid using CMAKE_PREFIX_PATH.

This should make it possible to use CMAKE_FIND_ROOT_PATH in a toolchain
file  without getting issues from the concatenation behavior, if it turn
out to be truly necessary for some toolchain.