Enable realtime lights for vertex-lit surfaces
Use the lightMapping GLSL shader instead of `generic` to render BSP
surfaces that are vertex-lit due to an explicit `rgbGen vertex`. This
makes them use the same code path as BSP surfaces where vertex lighting
is automatically selected due to the absence of a lightmap.

So now it is possible to render realtime lights on explicitly vertex-lit
surfaces, though they are still not as bright as they should be due to
https://github.com/DaemonEngine/Daemon/issues/1415.

In the following commit, this will let us remove some ugly code used for
applying the overbright factor to the `generic` shader.