Add r_forceBlendRegime: force linear/naive blending
Add the cvar r_forceBlendRegime which if set to 1 forces naive blending
even on a linear lighting map, or if set to 2 forces linear blending on
a non-linear map. Overbright factors are adjusted to produce a result as
similar as possible to the map's "native" mode, assuming tone mapping is
disabled.

Correctly displaying a linear rgbGen or linear colormap is not
implemented. Anyway there is already a console warning that should fire
in this scenario.