Fix cvar changes sometimes not writing autogen
A command such as /set common.framerate.max 99, when the cvar was
previously marked as archive, could fail to cause the configuration file
to be written out. This happened because the code relied on the flags
field of the C API cvar_t object to detect updates, but Cvar::AddFlags
failed to update cvar_t.

Note that setting cvars in the UI was not affected because the UI
re-adds the archive flag every time and the act of adding the flag
triggers an update.